Rename navigation-bar.d.ts to navigation-bar.ts
