+++ frontend/src/components/common/layout-components/NavigationBar/NavigationBar.tsx
@@ -22,7 +22,7 @@ import LogoButton from "./LogoButton";
 // Text to display in the hamburger menu, navbar and the corresponding link
 // -,- in the menu and the corresponding link
 
-import type { NavigationData } from "./navigation-bar.d.ts";
+import type { NavigationData } from "./navigation-bar";
 
 interface Props {
   window?: () => Window;
