+++ frontend/src/components/common/layout-components/NavigationBar/navigation-bar.ts
